Moved AnyEscapableString hierarchy into RoddyToolLib.
