+++ src/main/groovy/de/dkfz/roddy/execution/jobs/FakeBEJob.groovy
@@ -8,10 +8,9 @@ package de.dkfz.roddy.execution.jobs
 
 
 import de.dkfz.roddy.core.InfoObject
-import de.dkfz.roddy.execution.AnyEscapableString
 import groovy.transform.CompileStatic
 
-import static de.dkfz.roddy.execution.EscapableString.*
+import static de.dkfz.roddy.tools.EscapableString.*
 
 /**
  * Created by heinold on 27.02.17.
