+++ src/main/groovy/de/dkfz/roddy/execution/jobs/cluster/lsf/rest/RestSubmissionCommand.groovy
@@ -6,18 +6,14 @@
 
 package de.dkfz.roddy.execution.jobs.cluster.lsf.rest
 
-import de.dkfz.roddy.StringConstants
+
 import de.dkfz.roddy.config.JobLog
-import de.dkfz.roddy.execution.AnyEscapableString
+import de.dkfz.roddy.tools.AnyEscapableString
 import de.dkfz.roddy.execution.jobs.BEJobID
 import de.dkfz.roddy.execution.jobs.SubmissionCommand
-import de.dkfz.roddy.tools.BashUtils
-import de.dkfz.roddy.tools.shell.bash.Service
 import groovy.transform.CompileStatic
 import org.apache.http.Header
 import sun.reflect.generics.reflectiveObjects.NotImplementedException
-
-import static de.dkfz.roddy.StringConstants.EMPTY
 
 /**
  * This class is used to create REST commands
