+++ src/test/groovy/de/dkfz/roddy/execution/CommandSpec.groovy
@@ -1,9 +1,11 @@
 package de.dkfz.roddy.execution
 
+import de.dkfz.roddy.tools.AnyEscapableString
+import de.dkfz.roddy.tools.BashInterpreter
 import spock.lang.Specification
 
 import java.nio.file.Paths
-import static de.dkfz.roddy.execution.EscapableString.*
+import static de.dkfz.roddy.tools.EscapableString.*
 
 class CommandSpec extends Specification {
