+++ src/test/groovy/de/dkfz/roddy/execution/ExecutableSpec.groovy
@@ -3,7 +3,7 @@ package de.dkfz.roddy.execution
 import spock.lang.Specification
 
 import java.nio.file.Paths
-import static de.dkfz.roddy.execution.EscapableString.*
+import static de.dkfz.roddy.tools.EscapableString.*
 
 class ExecutableSpec extends Specification {
