+++ src/test/groovy/de/dkfz/roddy/execution/jobs/cluster/lsf/AbstractLSFJobManagerSpec.groovy
@@ -2,9 +2,9 @@ package de.dkfz.roddy.execution.jobs.cluster.lsf
 
 import com.google.common.collect.LinkedHashMultimap
 import de.dkfz.roddy.config.ResourceSet
-import de.dkfz.roddy.execution.AnyEscapableString
+import de.dkfz.roddy.tools.AnyEscapableString
 import de.dkfz.roddy.execution.BEExecutionService
-import de.dkfz.roddy.execution.BashInterpreter
+import de.dkfz.roddy.tools.BashInterpreter
 import de.dkfz.roddy.execution.io.ExecutionResult
 import de.dkfz.roddy.execution.jobs.Command
 import de.dkfz.roddy.execution.jobs.JobManagerOptionsBuilder
@@ -12,8 +12,6 @@ import spock.lang.Shared
 import spock.lang.Specification
 
 import java.time.Duration
-
-import static de.dkfz.roddy.execution.EscapableString.*
 
 class AbstractLSFJobManagerSpec extends Specification {
